requirements.txt
numpy==1.19.5
scipy==1.5.4
matplotlib==3.3.4